Major refactoring to organize methods
