--- seleniumbase/fixtures/tools.py
@@ -1,9 +1,0 @@
-"""
-This module imports all the commonly used fixtures in one place so that
-every test doesn't need to import a number of different fixtures.
-"""
-
-from seleniumbase.fixtures.page_actions import *  # noqa
-from seleniumbase.fixtures.page_utils import *  # noqa
-from seleniumbase.fixtures.errors import *  # noqa
-from seleniumbase.fixtures.xpath_to_css import *  # noqa
